SOA Lab2 service2

+++ SOA/Lab2/service2/src/main/java/dev/dto/Percentage.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Component;
 @Setter
 @Component
 public class Percentage {
-    Double value;
+    int value;
 }
